proxy support, part 1 - separate request building to another class
