+++ src/GeocodeSharp/GeocodeSharp.csproj
@@ -58,6 +58,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DefaultRequestBuilder.cs" />
     <Compile Include="Google\AddressComponent.cs" />
     <Compile Include="Google\ComponentFilter.cs" />
     <Compile Include="Google\GeocodeClient.cs" />
@@ -70,7 +71,9 @@
     <Compile Include="Google\IGeocodeClient.cs" />
     <Compile Include="Google\LocationType.cs" />
     <Compile Include="Google\UsageMode.cs" />
+    <Compile Include="NginxProxyRequestBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RequestBuilderBase.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="key.snk" />
